+++ ui/src/pages/namespace/Explorer/Endpoint/EndpointEditor/Form/plugins/Inbound/HeaderManipulationForm/HeaderArrayInput.tsx
@@ -1,4 +1,4 @@
-import { ArrayForm } from "~/components/Form/ArrayInput";
+import { ArrayForm } from "~/components/Form/Array";
 import { ControllerRenderProps } from "react-hook-form";
 import { HeaderManipulationFormSchemaType } from "../../../../schema/plugins/inbound/headerManipulation";
 import Input from "~/design/Input";
