+++ ui/src/pages/namespace/Explorer/Service/ServiceEditor/Form/EnvsArrayInput.tsx
@@ -1,4 +1,4 @@
-import { ArrayForm } from "~/components/Form/ArrayInput";
+import { ArrayForm } from "~/components/Form/Array";
 import { ControllerRenderProps } from "react-hook-form";
 import Input from "~/design/Input";
 import { ServiceFormSchemaType } from "../schema";
